A_Night_at_the_Museum
I came up with a very efficient solution for this problem, which is using the modulo operator. I am so happy to have come up with this solution as it took me a lot of mathematical work :) 
Basically, if you want to know the shortest path between a and z, 122- 97 = 25, which is the wrong solution. The right answer will be 97-122+26=1. The only problem here 
You can't use the same formula if we are working with a and b, 97-96+26=27, which is wrong, so we use the modulo to get the remainder of the calculation every time. 
Mar/25/2023 13:02UTC+2
15 ms
0 KB