Update A_Buy_a_Shovel.cpp
I came up with a way better approach here:
you calculate the full price and check if the change equals r or not as you have unlimited numbers of 10 coins and one coin of value r. so if the change equals r, "i" will be the minimum number of shovels.